requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.54
botocore==1.31.54
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-rest-passwordreset==1.3.0
django-storages==1.14
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
gunicorn==21.2.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.17
